OpenBSD 3.0

+++ openbsd/sys/netipx/ipx_cksum.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: ipx_cksum.c,v 1.2 1996/10/26 09:34:49 mickey Exp $	*/
+/*	$OpenBSD: ipx_cksum.c,v 1.4 2001/09/23 10:22:13 mickey Exp $	*/
 
 /*-
  *
@@ -103,9 +103,9 @@ ipx_cksum(m, len)
 		 * 32 bit arithmetic registers for adding,
 		 * with carries from the low added
 		 * into the high (by normal carry-chaining)
-		 * so long as we fold back before 16 carries have occured.
+		 * so long as we fold back before 16 carries have occurred.
 		 */
-		if (1 & (int) w)
+		if (1 & (long) w)
 			goto uuuuglyy;
 #ifndef TINY
 /* -DTINY reduces the size from 1250 to 550, but slows it down by 22% */
